+++ Mapping/ControllerConverters/ControllerConverter.cpp
@@ -1,4 +1,4 @@
 class ControllerConverter {
     public:
-        virtual contoller_t convert(contoller_t input_controller, string user_config_path) = 0;
-}
+        virtual controller_t convert(controller_t input_controller, string user_config_path) = 0;
+};
