+++ Mapping/ControllerConverters/ControllerConverter.h
@@ -1,8 +1,4 @@
-#include "../ControllerModels/Controller.h"
-#include "../ControllerModels/XboxController.h"
-#include "../ControllerModels/SnesController.h"
-
 class ControllerConverter {
     public:
-        virtual contoller_t convert(contoller_t input_controller, string user_config_path);
+        virtual controller_t convert(controller_t input_controller, string user_config_path);
 };
