+++ Mapping/ControllerConverters/ControllerConverterFactory.h
@@ -1,7 +1,5 @@
-#include "ControllerConverter.h"
-
 class ControllerConverterFactory
 {
  public:
-  static ControllerConverter createConverter(controller_t input_controller, controller_t output_controller);
+  static ControllerConverter* createConverter(controller_t input_controller, controller_t output_controller);
 };
