+++ Mapping/ControllerModels/SnesController.h
@@ -1,4 +1,4 @@
-struct snes_controller_t: contoller_t {
+struct snes_controller_t: controller_t {
 	bool A;
 	bool B;
 	bool X;
